+++ kkanbu/board/urls.py
@@ -1,9 +1,15 @@
+from django.conf.urls import url
 from rest_framework.routers import DefaultRouter
 
-from .views import CommentViewSet, PostViewSet
+from .views import BoardListView, CommentViewSet  # PostViewSet,
 
 router = DefaultRouter()
-router.register(r"post", PostViewSet)
+# router.register(r"post", PostViewSet)
 router.register(r"comment", CommentViewSet)
 
-urlpatterns = router.urls
+app_name = "board"
+urlpatterns = [
+    url(r"^$", BoardListView.as_view(), name="board-list"),
+]
+
+urlpatterns += router.urls